made safe for nonexistent output file

+++ reset.py
@@ -1,6 +1,7 @@
 import os
 
-os.remove('nick_usage.csv')
+if os.path.exists('nick_usage.csv'):
+    os.remove('nick_usage.csv')
 
 file = open('nick_usage.csv', 'w')
 file.write('pull_date,pull_time,location,update_time,pct_full,date,time,wday\n')
